+++ src/components/table/infinity-table-empty.tsx
@@ -3,7 +3,7 @@ import { faMagnifyingGlass } from "@fortawesome/free-solid-svg-icons";
 import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import { EmptyState } from "../ui/empty-state";
 
-function PaginatedTableEmpty() {
+function InfinityTableEmpty() {
   return (
     <Table.Row bgColor={"transparent"}>
       <Table.Cell colSpan={8} minWidth={"100%"}>
@@ -22,4 +22,4 @@ function PaginatedTableEmpty() {
   );
 }
 
-export default PaginatedTableEmpty;
+export default InfinityTableEmpty;